refactor(utils): move `formatters` to outer scope
