+++ app/profile/components/common/selects/select-view.tsx
@@ -11,7 +11,7 @@ import {
   SelectValue,
 } from '@app/components/ui/select'
 import { View } from '@app/types'
-import { formatSearchParams } from '@app/profile/utils/formatters'
+import { formatSearchParams } from '@app/utils/formatters'
 
 export interface SelectViewProps {
   initialValue: View
