+++ app/utils/formatters/artists.spec.ts
