+++ app/utils/formatters/search-params.spec.ts
